Store configuration inside config class
