+++ robotidy/api.py
@@ -5,49 +5,50 @@ from typing import Optional
 
 from robotidy.app import Robotidy
 from robotidy.cli import TransformType, find_and_read_config, validate_regex
+from robotidy.config import Config
 from robotidy.disablers import RegisterDisablers
 from robotidy.files import DEFAULT_EXCLUDES
 from robotidy.utils import ROBOT_VERSION, GlobalFormattingConfig
 
 
-class RobotidyAPI(Robotidy):
-    def __init__(self, src: str, output: Optional[str], **kwargs):
-        config = find_and_read_config((src,))
-        config = {k: str(v) if not isinstance(v, (list, dict)) else v for k, v in config.items()}
-        converter = TransformType()
-        transformers = [converter.convert(tr, None, None) for tr in config.get("transform", ())]
-        configurations = [converter.convert(c, None, None) for c in config.get("configure", ())]
-        space_count = kwargs.get("spacecount", None) or int(config.get("spacecount", 4))
-        formatting_config = GlobalFormattingConfig(
-            space_count=space_count,
-            indent=kwargs.get("indent", None) or int(config.get("indent", space_count)),
-            separator=kwargs.get("separator", None) or config.get("separator", "space"),
-            line_sep=config.get("lineseparator", "native"),
-            start_line=kwargs.get("startline", None) or int(config["startline"]) if "startline" in config else None,
-            end_line=kwargs.get("endline", None) or int(config["endline"]) if "endline" in config else None,
-            line_length=kwargs.get("line_length", None) or int(config.get("line_length", 120)),
-        )
-        exclude = config.get("exclude", None)
-        extend_exclude = config.get("extend_exclude", None)
-        exclude = validate_regex(exclude if exclude is not None else DEFAULT_EXCLUDES)
-        extend_exclude = validate_regex(extend_exclude)
-        super().__init__(
-            transformers=transformers,
-            transformers_config=configurations,
-            src=(),
-            exclude=exclude,
-            extend_exclude=extend_exclude,
-            skip_gitignore=False,
-            overwrite=False,
-            show_diff=False,
-            formatting_config=formatting_config,
-            verbose=False,
-            check=False,
-            output=output,
-            force_order=False,
-            target_version=ROBOT_VERSION.major,
-            color=False,
-        )
+def get_robotidy(src: str, output: Optional[str], **kwargs):
+    config = find_and_read_config((src,))
+    config = {k: str(v) if not isinstance(v, (list, dict)) else v for k, v in config.items()}
+    converter = TransformType()
+    transformers = [converter.convert(tr, None, None) for tr in config.get("transform", ())]
+    configurations = [converter.convert(c, None, None) for c in config.get("configure", ())]
+    space_count = kwargs.get("spacecount", None) or int(config.get("spacecount", 4))
+    formatting_config = GlobalFormattingConfig(
+        space_count=space_count,
+        indent=kwargs.get("indent", None) or int(config.get("indent", space_count)),
+        separator=kwargs.get("separator", None) or config.get("separator", "space"),
+        line_sep=config.get("lineseparator", "native"),
+        start_line=kwargs.get("startline", None) or int(config["startline"]) if "startline" in config else None,
+        end_line=kwargs.get("endline", None) or int(config["endline"]) if "endline" in config else None,
+        line_length=kwargs.get("line_length", None) or int(config.get("line_length", 120)),
+    )
+    exclude = config.get("exclude", None)
+    extend_exclude = config.get("extend_exclude", None)
+    exclude = validate_regex(exclude if exclude is not None else DEFAULT_EXCLUDES)
+    extend_exclude = validate_regex(extend_exclude)
+    configuration = Config(
+        transformers=transformers,
+        transformers_config=configurations,
+        src=(),
+        exclude=exclude,
+        extend_exclude=extend_exclude,
+        skip_gitignore=False,
+        overwrite=False,
+        show_diff=False,
+        formatting=formatting_config,
+        verbose=False,
+        check=False,
+        output=output,
+        force_order=False,
+        target_version=ROBOT_VERSION.major,
+        color=False,
+    )
+    return Robotidy(config=configuration)
 
 
 def transform_model(model, root_dir: str, output: Optional[str] = None, **kwargs) -> Optional[str]:
@@ -60,14 +61,14 @@ def transform_model(model, root_dir: str, output: Optional[str] = None, **kwargs
     such as ``spacecount``, ``startline`` and ``endline``.
     :return: The transformed model converted to string or None if no transformation took place.
     """
-    transformer = RobotidyAPI(root_dir, output, **kwargs)
+    robotidy_class = get_robotidy(root_dir, output, **kwargs)
     disabler_finder = RegisterDisablers(
-        transformer.formatting_config.start_line, transformer.formatting_config.end_line
+        robotidy_class.config.formatting.start_line, robotidy_class.config.formatting.end_line
     )
     disabler_finder.visit(model)
     if disabler_finder.file_disabled:
         return None
-    diff, _, new_model = transformer.transform(model, disabler_finder.disablers)
+    diff, _, new_model = robotidy_class.transform(model, disabler_finder.disablers)
     if not diff:
         return None
     return new_model.text
